requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
